added admin privilages

+++ routes/products.js
@@ -13,23 +13,38 @@ router.get("/:id?", (req, res) => {
 })
 
 router.post("/", (req, res) => {
-  const product = new Product(req.body)
-  res.send(products.save(product))
+  const admin = req.headers.admin
+  if (admin) {
+    const product = new Product(req.body)
+    res.send(products.save(product))
+  } else {
+    res.send("No tiene permisos para esta ruta")
+  }
 })
 
 router.put("/:id", (req, res) => {
-  const id = req.params.id
-  const updatedProduct = new Product(req.body)
-  const idx = products.find((p) => p.id === id)
-  if (idx != -1) {
-    products[idx] = updatedProduct
-    res.json({ actualizado: products[idx] })
-  } else res.json("Id no encontrado")
+  const admin = req.headers.admin
+  if (admin) {
+    const id = req.params.id
+    const updatedProduct = new Product(req.body)
+    const idx = products.find((p) => p.id === id)
+    if (idx != -1) {
+      products[idx] = updatedProduct
+      res.json({ actualizado: products[idx] })
+    } else res.send("Id no encontrado")
+  } else {
+    res.send("No tiene permisos para esta ruta")
+  }
 })
 
 router.delete("/:id", (req, res) => {
-  const id = req.params.id
-  res.send(products.deleteById(id))
+  const admin = req.headers.admin
+  if (admin) {
+    const id = req.params.id
+    res.send(products.deleteById(id))
+  } else {
+    res.send("No tiene permisos para esta ruta")
+  }
 })
 
 class Product {
